Port classification time calculation from ActiveData to sql.telemetry.mozilla.org
The old code was at
https://github.com/Archaeopteryx/active-data-recipes/tree/classification_time-simple

The ActiveData source is currently broken (bug 1688718) and scheduled to be
decommissioned (bug 1692453). This script switches to
https://sql.telemetry.mozilla.org Due to its limitations, it's not possible
anymore to judge a task as waiting for its rerun.

The classification time is the time between the CI task completing with a result
of failure ('busted' and 'testfailed' in Treeherder) and the result getting
classified as e.g. intermittent failure and getting a bug number assigned for
the issue.